refactor(product-components): extract StepCard into product-components
Move StepCard from WipeDeviceModal into @trezor/product-components
so it can be reused by both WipeDeviceModal and OsAndTrezorCleanupModal.
