+++ packages/suite/src/views/settings/SettingsDevice/WipeDevice/WipeDeviceModal.tsx
@@ -1,81 +1,19 @@
-import { type ReactNode, useState } from 'react';
+import { useState } from 'react';
 
 import { isFulfilled } from '@reduxjs/toolkit';
 
 import { Translation } from '@suite/intl';
 import { events } from '@suite-common/analytics';
 import { wipeDeviceThunk } from '@suite-common/wallet-core';
-import {
-    Button,
-    Card,
-    Column,
-    Divider,
-    Icon,
-    type IconName,
-    Modal,
-    Paragraph,
-    Row,
-    Text,
-} from '@trezor/components';
+import { Button, Column, Modal } from '@trezor/components';
 import { isDeviceInBootloaderMode } from '@trezor/device-utils';
+import { StepCard } from '@trezor/product-components';
 
 import { useDevice, useDispatch } from 'src/hooks/suite';
 import { useAnalytics } from 'src/support/useAnalytics';
 
 type WipeDeviceModalProps = {
     onCancel: () => void;
-};
-
-type StepCardProps = {
-    heading: ReactNode;
-    description: ReactNode;
-    actions: ReactNode;
-    icon: IconName;
-    descriptionTypographyStyle?: 'body-md-strong' | 'inherit';
-    state: 'default' | 'confirmed' | 'pending';
-};
-
-export const StepCard = ({
-    heading,
-    description,
-    actions,
-    icon,
-    state,
-    descriptionTypographyStyle = 'body-md-strong',
-}: StepCardProps) => {
-    const iconIntent = state === 'confirmed' ? 'brand' : 'neutral';
-    const iconPriority = state === 'confirmed' ? 'primary' : 'secondary';
-    const textIntent = state === 'confirmed' ? 'brand' : 'neutral';
-    const textPriority = state === 'confirmed' ? 'primary' : 'secondary';
-
-    return (
-        <Card paddingType="none" fillType={state === 'pending' ? 'flat' : 'default'}>
-            <Column>
-                <Row gap={8} padding={{ horizontal: 16, vertical: 12 }}>
-                    <Icon
-                        name={state === 'confirmed' ? 'check' : icon}
-                        intent={iconIntent}
-                        priority={iconPriority}
-                        size={20}
-                    />
-                    <Text typographyStyle="body-sm" intent={textIntent} priority={textPriority}>
-                        {heading}
-                    </Text>
-                </Row>
-                {state === 'default' && (
-                    <>
-                        <Divider margin={0} />
-                        <Column gap={16} padding={{ horizontal: 16, vertical: 12 }}>
-                            <Paragraph typographyStyle={descriptionTypographyStyle}>
-                                {description}
-                            </Paragraph>
-                            <Row gap={12}>{actions}</Row>
-                        </Column>
-                    </>
-                )}
-            </Column>
-        </Card>
-    );
 };
 
 export const WipeDeviceModal = ({ onCancel }: WipeDeviceModalProps) => {
